% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fBallInterceptionProbability.R
\name{fBallInterceptionProbability}
\alias{fBallInterceptionProbability}
\title{Pitch control}
\usage{
fBallInterceptionProbability(tti_sigma, time_to_intercept, Time)
}
\description{
Pitch control
}
