% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fBeeswarmChart.R
\name{fBeeswarmChart}
\alias{fBeeswarmChart}
\title{Those charts with normailsed metric values and a bright mark indicating
where the PlayerName of interest lies}
\usage{
fBeeswarmChart(
  dtPlayerMetrics,
  vcColumnsToIndex,
  dtMetricCategorisation,
  iPlayerId,
  cTitle,
  cFontFamily = "arial",
  cForegroundColour = "red",
  cBackgroundColour = "black",
  cFontColour = "white",
  vnExpand = c(0.1, 0.1)
)
}
\arguments{
\item{dtPlayerMetrics}{A dataset with one row for each PlayerName, and various
metrics about the PlayerName declared in separate columns. Refer to the
dtPlayerMetrics dataset packaged with the library for an example}

\item{vcColumnsToIndex}{The non-metric columns in your dataset, these are
typically columns like name, age, team, position, etc.}

\item{dtMetricCategorisation}{A table with metadata about the variables in
dtPlayerMetrics. Refer to the dtMetricCategorisation object declared in the
library for an example.}

\item{iPlayerId}{The ID of the player you want visualised}

\item{cTitle}{The title on the chart}
}
\description{
Those charts with normailsed metric values and a bright mark indicating
where the PlayerName of interest lies
}
\examples{
fBeeswarmChart (
   dtPlayerMetrics,
   vcColumnsToIndex = c('playerId','PlayerName','TeamName'),
   dtMetricCategorisation,
   iPlayerId = 2,
   cTitle = 'Sample'
)
}
