% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCalculatePolygonCoordinate.R
\name{fCalculatePolygonCoordinate}
\alias{fCalculatePolygonCoordinate}
\title{Draws a block based on the angle bucket and length bucket, respectively 
indicated by x and y, used by fPlotSonar}
\usage{
fCalculatePolygonCoordinate(nAngleMin_rad, nAngleMax_rad, nRadiusMin,
  nRadiusMax, fTrigFunction)
}
\description{
Hard to describe it. Play with the example
}
\examples{
plot(
   fCalculatePolygonCoordinate(
      nAngleMin_rad = pi/4,
      nAngleMax_rad = pi/2,
      nRadiusMin = 1,
      nRadiusMax = 2,
      cos   
   ),
   fCalculatePolygonCoordinate(
      nAngleMin_rad = pi/4,
      nAngleMax_rad = pi/2,
      nRadiusMin = 1,
      nRadiusMax = 2,
      sin   
   )      
)
}
