% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fConvertTrackingDataWideToLong.R
\name{fConvertTrackingDataWideToLong}
\alias{fConvertTrackingDataWideToLong}
\title{Converts tracking data from wide to long}
\usage{
fConvertTrackingDataWideToLong(dtData)
}
\description{
Very tightly tied to how https://github.com/metrica-sports/sample-data is
organised right now since I don't know how tracking data is otherwise
organised
}
