% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fDrawVoronoiFromVoronoiCoordinates.R
\name{fDrawVoronoiFromVoronoiCoordinates}
\alias{fDrawVoronoiFromVoronoiCoordinates}
\title{Draws the Voronoi plot from the Voronoi coordinates}
\usage{
fDrawVoronoiFromVoronoiCoordinates(
  dtVoronoiCoordinates,
  dtTrackingSlice,
  nXLimit,
  nYLimit,
  markFutureTrajectoryFor,
  dtTrackingSliceFuture
)
}
\description{
Draws the Voronoi plot from the Voronoi coordinates
}
