% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetBlocksThroughWhichPassesWent.R
\name{fGetBlocksThroughWhichPassesWent}
\alias{fGetBlocksThroughWhichPassesWent}
\title{Identifies the blocks in a grid through which a pass (
or more generally a line segment ) goes}
\usage{
fGetBlocksThroughWhichPassesWent(
  dt_passes,
  x_lim = c(-60, 60),
  y_lim = c(-40, 40),
  x_gap = 5,
  y_gap = 5
)
}
\arguments{
\item{dt_passes}{a data.table with x, y, xend, yend, and optionally id}

\item{x_lim}{the span of x coordinates on your grid - c(min x, max x)}

\item{y_lim}{the span of x coordinates on your grid - c(min y, max y)}

\item{x_gap}{the spacing between blocks of the grid on the x dim}

\item{y_gap}{the spacing between blocks of the grid on the y dim}
}
\description{
Identifies the blocks in a grid through which a pass (
or more generally a line segment ) goes
}
\examples{

library(data.table)
dt_passes = rbind(
   data.table(x = -27, y = 7, xend = -37, yend = 16, id = 1),
   data.table(x = 27, y = 7, xend = 37, yend = 16, id = 2),
   data.table(x = 27, y = -7, xend = 37, yend = -16, id = 3),
   data.table(x = -27, y = -7, xend = -37, yend = -16, id = 4),
   data.table(x = -25, y = 25, xend = -10, yend = 35, id = 5),
   data.table(x = 25, y = 25, xend = 10, yend = 35, id = 6),
   data.table(x = 25, y = -25, xend = 10, yend = -35, id = 7),
   data.table(x = -25, y = -25, xend = -10, yend = -35, id = 8),
   data.table(x = -8, y = -20, xend = 8, yend = -20, id = 9),
   data.table(x = 0, y = 0, xend = 0, yend = 20, id = 10)
)
dt_grids_pass = fGetBlocksThroughWhichPassesWent(
   dt_passes = dt_passes,
   x_lim = c(-60,60),
   y_lim = c(-40,40),
   x_gap = 5,
   y_gap = 5
)


library(ggplot2)
ggplot() +
   geom_pitch(nXStart = -60, nYStart = -40, nXEnd = 60, nYEnd = 40) +
   geom_tile(
      data = dt_grids_pass,
      aes(x = block_x, y = block_y),
      fill = 'cyan',
      color = 'black'
   ) +
   geom_segment(
      data = dt_passes,
      aes(x = x, y = y, xend = xend, yend = yend),
      color = 'red',
      arrow=arrow(type="closed",ends="last",length=unit(0.1,"cm"))
   )

}
