% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetPitchCoordinates.R
\name{fGetPitchCoordinates}
\alias{fGetPitchCoordinates}
\title{Pitch dimension coordinates}
\usage{
fGetPitchCoordinates(
  nXStart = -60,
  nYStart = -40,
  nXEnd = 60,
  nYEnd = 40,
  iXStripes = 0,
  iYStripes = 0
)
}
\description{
Pitch dimension coordinates
}
