% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetPitchDimensions.R
\name{fGetPitchDimensions}
\alias{fGetPitchDimensions}
\title{Pitch dimension coordinates}
\usage{
fGetPitchDimensions(nXStart = 0, nYStart = 0, nXEnd = 120, nYEnd = 80)
}
\description{
Pitch dimension coordinates
}
