% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetPitchTransformedCoordinates.R
\name{fGetPitchTransformedCoordinates}
\alias{fGetPitchTransformedCoordinates}
\title{Uses POV and fGetPitchCoordinates to give reoriented xy coords of pitch for plotting}
\usage{
fGetPitchTransformedCoordinates(
  lPitchCoordinates,
  mOriginCoordinates = NULL,
  mScreenCoordinates = NULL,
  mViewBeginsFromCoordinates = NULL,
  mVectorPointingUpOnScreen = cbind(0, 0, 1)
)
}
\description{
Uses POV and fGetPitchCoordinates to give reoriented xy coords of pitch for plotting
}
