% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetVoronoiFromDataTable.R
\name{fGetVoronoiFromDataTable}
\alias{fGetVoronoiFromDataTable}
\title{Pass network ( WIP )}
\usage{
fGetVoronoiFromDataTable(
  dtTrackingSlice,
  xMinBB = 0,
  yMinBB = 0,
  xMaxBB = 1,
  yMaxBB = 1
)
}
\description{
Plots a marker for each player at their median passing position, and draws
connections between players to represent the number of passes exchanged
between them
}
