% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetVoronoiFromTable.R
\name{fGetVoronoiFromTable}
\alias{fGetVoronoiFromTable}
\title{Pass network ( WIP )}
\usage{
fGetVoronoiFromTable(dtFrame, nXLimit = 120, nYlimit = 80)
}
\description{
Plots a marker for each player at their median passing position, and draws
connections between players to represent the number of passes exchanged
between them
}
