% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetVoronoi.R
\name{fGetVoronoiFromVectors}
\alias{fGetVoronoiFromVectors}
\title{Pass network ( WIP )}
\usage{
fGetVoronoiFromVectors(
  x = NULL,
  y = NULL,
  xMinBB = NULL,
  yMinBB = NULL,
  xMaxBB = NULL,
  yMaxBB = NULL
)
}
\description{
Plots a marker for each player at their median passing position, and draws
connections between players to represent the number of passes exchanged
between them
}
