% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fJsonToListOfTables.R
\name{fJsonToListOfTables}
\alias{fJsonToListOfTables}
\title{Primarily to convert statsbomb files to a list of data.tables}
\usage{
fJsonToListOfTables(lMatchJson, vcTagsToExtract, nXLimit = 120, nYLimit = 80)
}
\arguments{
\item{lMatchJson}{refer example}

\item{vcTagsToExtract}{refer example}

\item{nXLimit}{Length of the pitch}

\item{nYLimit}{Breadth of the pitch}
}
\description{
The function has been named generically so that tomorrow it can be used 
with other json syntaxes as well. As of right now, it has been coded up
in a manner that it works with StatsBomb data only.
}
\examples{
# read in the raw json as a list
lMatchJson = rjson::fromJSON(file = cMatchFileName)

# get one table each for the starting XI, passes, and shots
lEvents = fJsonToListOfTables (
   lMatchJson = lMatchJson,
   vcTagsToExtract = c('Starting XI','Pass','Shot')
)
}
