% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fJsonToTabular.R
\name{fJsonToTabular}
\alias{fJsonToTabular}
\title{Helper function to fJsonToListOfTables}
\usage{
fJsonToTabular(lEvent)
}
\description{
Breaks individual json chunks into tables, takes care of duplicate 
column names.
}
