% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fParseTrackingData.R
\name{fParseTrackingData}
\alias{fParseTrackingData}
\title{Pass network ( WIP )}
\usage{
fParseTrackingData(cFilePath, cTag = "")
}
\description{
Plots a marker for each player at their median passing position, and draws
connections between players to represent the number of passes exchanged
between them
}
