% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fParseTrackingDataBothTeams.R
\name{fParseTrackingDataBothTeams}
\alias{fParseTrackingDataBothTeams}
\title{Gives cleanly parsed data}
\usage{
fParseTrackingDataBothTeams(
  cRootPath,
  cGameName,
  bOutputLong,
  nXLimit = 120,
  nYLimit = 80,
  xMaxBB = 1,
  yMaxBB = 1,
  nUpperLimitSpeed = 10
)
}
\description{
Very tightly tied to how https://github.com/metrica-sports/sample-data is
organised right now since I don't know how tracking data is otherwise
organised
}
