% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fParseTrackingDataOneTeam.R
\name{fParseTrackingDataOneTeam}
\alias{fParseTrackingDataOneTeam}
\title{Pass network ( WIP )}
\usage{
fParseTrackingDataOneTeam(
  cFilePath,
  cTag = "",
  nXLimit = 120,
  nYLimit = 80,
  xMaxBB = 1,
  yMaxBB = 1
)
}
\description{
Plots a marker for each player at their median passing position, and draws
connections between players to represent the number of passes exchanged
between them
}
