% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPassNetworkChart.R
\name{fPassNetworkChart}
\alias{fPassNetworkChart}
\title{Pass network ( WIP )}
\usage{
fPassNetworkChart(dtPasses, dtPlayerLabels, nXLimit = 120,
  nYLimit = 80, nSegmentWidth = 1)
}
\arguments{
\item{dtPasses}{a data.table with the columns playerId ( the player who
made the pass, ) recipientPlayerId ( the player who received the pass, )
Success ( 1/0 for whether the pass reached the recipient, ) x ( the 
coordinate along the length of the pitch, 0 is defensive end, nXLimit is 
offensive end, ) and y ( along the breadth of the pitch, 0 is right wing and 
nYLimit is left wing. ) Refer to the dtPasses dataset packaged with the 
library}

\item{dtPlayerLabels}{a data.table with the colums playerId ( same as 
dtPasses, ) and playerName ( the label that the point of the respective 
player should be labelled as. ) Leaving this blank will mean no labels in 
the diagram. Refer to the dtPlayerLabels dataset packaged with the library.}

\item{nXLimit}{Length of the pitch}

\item{nYLimit}{Breadth of the pitch}
}
\description{
Plots a marker for each player at their median passing position, and draws
connections between players to represent the number of passes exchanged 
between them
}
\examples{
fPassNetworkChart(
   dtPasses,
   dtPlayerLabels
)
}
