% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPlotPitchControl.R
\name{fPlotPitchControl}
\alias{fPlotPitchControl}
\title{Pitch control support function}
\usage{
fPlotPitchControl(
  lPitchControl,
  cFolderPathToSaveImagesSubdir = NULL,
  DelayBetweenFrames = 10,
  nXLimit = 120,
  nYLimit = 80,
  vcColourAssignment = c(Home = "red", Away = "blue", Ball = "black")
)
}
\description{
Pitch control support function
}
