% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPlotSonar.R
\name{fPlotSonar}
\alias{fPlotSonar}
\title{This function has only been tested with 120 X 80 dimension dataset, in the 
two ways shown in the charts section. Any deviation from those may 
produce odd results.}
\usage{
fPlotSonar(dtPassesToPlot, iBlocksInFirstRing = 4, iNbrRings = 8,
  nZoomFactor = NULL, nLocation2Size = 80, nLocation1Size = 120,
  bAddPitchBackground = F, cTitle = NULL)
}
\description{
This function has only been tested with 120 X 80 dimension dataset, in the 
two ways shown in the charts section. Any deviation from those may 
produce odd results.
}
