% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fSimpleTimeToIntercept.R
\name{fSimpleTimeToIntercept}
\alias{fSimpleTimeToIntercept}
\title{Pitch control support function}
\usage{
fSimpleTimeToIntercept(
  reaction_time,
  VelocityX,
  VelocityY,
  position_x,
  position_y,
  vmax,
  target_x,
  target_y
)
}
\description{
Pitch control support function
}
