% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pitch.R
\name{geom_pitch}
\alias{geom_pitch}
\title{Pitch background}
\usage{
geom_pitch(
  nXStart = 0,
  nYStart = 0,
  nXEnd = 120,
  nYEnd = 80,
  mOrigin = NULL,
  mScreenCoordinate = NULL,
  cLineColour = "#BBBBBB",
  cPitchColour = "#038253"
)
}
\arguments{
\item{nXStart}{Start x coordinate of the pitch}

\item{nYStart}{Start y coordinate of the pitch}

\item{nXEnd}{End x coordinate of the pitch}

\item{nYEnd}{Start y coordinate of the pitch}

\item{cLineColour}{the colour of the line markings on the pitch}

\item{cPitchColour}{the colour of the grass on the pitch}

\item{plotObject}{a ggplot object. Can just be a blank ggplot object too}
}
\description{
Adds pitch markings, green background, etc. for plot background to look like
a football pitch. ggplot adds things in order so initiate a ggplot object,
add the pitch marking with this function, and then add the data you want to
add
}
\examples{
ggplot() + geom_pitch()
}
