% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pitch_from_coords_obj.R
\name{geom_pitch_from_coords_obj}
\alias{geom_pitch_from_coords_obj}
\title{part of the geom_pitch backend}
\usage{
geom_pitch_from_coords_obj(
  lPitchCoordinates,
  cLineColour = "#BBBBBB",
  cPitchColour = "#038253",
  nSize = 1,
  vcToPlot = c("Markings", "Stripes", "Goalframe", "Goalnet")
)
}
\description{
part of the geom_pitch backend
}
