% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_likelihood.R
\name{get_likelihood}
\alias{get_likelihood}
\title{Likelihood of number of success of an arbitrary set of probabilities}
\usage{
get_likelihood(probs, n_buckets = Inf)
}
\arguments{
\item{probs}{a vector of the probabilities of all the independent events}

\item{n_buckets}{the number of buckets to break the 0 to 1 probability
space in and combine events falling within that bucket to speed up the
processing.}
}
\description{
Is just binomial distributions under the hood
}
\examples{
dt_success_probs = get_likelihood(
   # probs = runif(100),
   probs = rep(.5,100),
   n_buckets = Inf
)
dt_success_probs[,plot(successes,pdf)]

}
