% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pitch.R
\name{theme_pitch}
\alias{theme_pitch}
\title{Lipstick for the pitch background plots
Refer to fAddPitchLines}
\usage{
theme_pitch()
}
\description{
Lipstick for the pitch background plots
Refer to fAddPitchLines
}
