% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fAddPitchLines.R
\name{fAddPitchLines}
\alias{fAddPitchLines}
\title{Adds pitch markings, green background, etc. for plot background to look like
a football pitch.}
\usage{
fAddPitchLines(plotObject, nXLimit = 105, nYLimit = 68,
  cLineColour = "white", cPitchColour = "#038253")
}
\description{
Adds pitch markings, green background, etc. for plot background to look like
a football pitch.
}
