% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCalculatePolygonCoordinate.R
\name{fCalculatePolygonCoordinate}
\alias{fCalculatePolygonCoordinate}
\title{Draws a block based on the angle bucket and length bucket, respectively 
indicated by x and y}
\usage{
fCalculatePolygonCoordinate(xmin2, xmax2, ymin2, ymax2, fTrigFunction)
}
\description{
Draws a block based on the angle bucket and length bucket, respectively 
indicated by x and y
}
