% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fDrawVoronoiFromTable.R
\name{fDrawVoronoiFromTable}
\alias{fDrawVoronoiFromTable}
\title{Draws a Voronoi diagram for a time slice}
\usage{
fDrawVoronoiFromTable(
  dtTrackingSlice,
  nXLimit,
  nYLimit,
  UseOneFrameEvery = 1,
  DelayBetweenFrames = 5,
  markTrajectoryFor = NULL,
  markOffsideLineFor = NULL,
  markLineBetween = NULL
)
}
\description{
Draws a Voronoi diagram for a time slice
}
