% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fDrawVoronoiFromTable.R
\name{fDrawVoronoiFromTable}
\alias{fDrawVoronoiFromTable}
\title{Draws a Voronoi diagram for a time slice}
\usage{
fDrawVoronoiFromTable(
  dtTrackingSlice,
  nXLimit = 120,
  nYlimit = 80,
  UseOneFrameEvery = 1,
  DelayBetweenFrames = 5,
  suppressWarnings = F
)
}
\description{
Draws a Voronoi diagram for a time slice
}
