% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetPitchControlProbabilities.R
\name{fGetPitchControlProbabilities}
\alias{fGetPitchControlProbabilities}
\title{Pitch control}
\usage{
fGetPitchControlProbabilities(
  lData,
  viTrackingFrame,
  params = c(),
  nYLimit = 120,
  nXLimit = 80,
  iGridCellsX = 60,
  bGetPlayerProbabilities = F
)
}
\description{
Pitch control
}
