% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetPitchControlProbabilities.R
\name{fGetPitchControlProbabilities}
\alias{fGetPitchControlProbabilities}
\title{Pitch control}
\usage{
fGetPitchControlProbabilities(
  lData,
  viTrackingFrame,
  params = c(),
  nYSpan = 80,
  nXSpan = 120,
  iGridCellsX = 120,
  bGetPlayerProbabilities = F,
  bVerbose = F,
  vnXArray = NULL,
  vnYArray = NULL
)
}
\arguments{
\item{lData}{output of fParseTrackingDataBothTeams}

\item{viTrackingFrame}{frames you want processed}

\item{params}{has defaults in the function code but refer to the function
code to see what values are there that you can override}

\item{nYSpan}{pitch dimensions, sideline to sideline, centre circle is 0,0}

\item{nXSpan}{pitch dimensions goal to goal, centre circle is 0,0}

\item{iGridCellsX}{resolution of the calculation}

\item{bGetPlayerProbabilities}{Whether only the overall probability per team
should be returned or even the deatils of the break up between players
should be returned}
}
\description{
Pitch control
}
