% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGetPitchDimensions.R
\name{fGetPitchDimensions}
\alias{fGetPitchDimensions}
\title{Pitch dimension coordinates}
\usage{
fGetPitchDimensions(nXStart = -60, nYStart = -40, nXEnd = 60, nYEnd = 40)
}
\description{
Pitch dimension coordinates
}
