% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPercentileBarChart.R
\name{fPercentileBarChart}
\alias{fPercentileBarChart}
\title{Ashwin Raman's percentile bar charts}
\usage{
fPercentileBarChart(dtDataset, vcColumnsToIndex, dtMetricCategorisation,
  iPlayerId, cTitle = NULL, nColumnWidthByTwo = 0.25,
  nBufferForTextHorizontal = 0.005, nBufferForTextVertical = 0.05,
  vnQuantileMarkers = NULL, bAddAbsoluteIndicator = F)
}
\arguments{
\item{vcColumnsToIndex}{The non-metric columns in your dataset, these are
typically columns like name, age, team, position, etc.}

\item{dtMetricCategorisation}{A table with metadata about the variables in
dtPlayerMetrics. Refer to the dtMetricCategorisation object declared in the
library for an example.}

\item{iPlayerId}{The ID of the player you want visualised}

\item{cTitle}{The title on the chart}

\item{bAddAbsoluteIndicator}{Percentiles can be a little misleading if the
underlying numbers aren't uniformly distributed. You can use the vertical
dimension to add annotations for an indicator of the absolute spread of the
values and where this particular player's values fall within that spread.}

\item{dtPlayerMetrics}{A dataset with one row for each PlayerName, and various
metrics about the PlayerName declared in separate columns. Refer to the
dtPlayerMetrics dataset packaged with the library for an example}
}
\description{
Each bar denotes the percentile the player is for that respective stat.
This is a much better alternative to radar plots.
}
\examples{
fPercentileBarChart(
   dtDataset = dtPlayerMetrics,
   vcColumnsToIndex = c('playerId','PlayerName','TeamName'),
   dtMetricCategorisation,
   iPlayerId = 2,
   cTitle = 'Sample'
)
}
