% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPlotPitchControl.R
\name{fPlotPitchControl}
\alias{fPlotPitchControl}
\title{Pitch control support function}
\usage{
fPlotPitchControl(
  lPitchControl,
  cFolderPathToSaveImagesSubdir = NULL,
  DelayBetweenFrames = 10
)
}
\description{
Pitch control support function
}
