% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pitch.R
\name{geom_pitch}
\alias{geom_pitch}
\title{Pitch background}
\usage{
geom_pitch(
  nXStart = -60,
  nYStart = -40,
  nXEnd = 60,
  nYEnd = 40,
  mOriginCoordinates = NULL,
  mScreenCoordinates = NULL,
  mViewBeginsFromCoordinates = NULL,
  mVectorPointingUpOnScreen = cbind(0, 0, 1),
  cLineColour = "#BBBBBB",
  cPitchColour = "#038253",
  vcToPlot = c("Markings", "Stripes", "Goalframe", "Goalnet"),
  iXStripes = 12,
  iYStripes = 8,
  nSize = 1
)
}
\arguments{
\item{nXStart}{Start x coordinate of the pitch}

\item{nYStart}{Start y coordinate of the pitch}

\item{nXEnd}{End x coordinate of the pitch}

\item{nYEnd}{Start y coordinate of the pitch}

\item{cLineColour}{the colour of the line markings on the pitch}

\item{cPitchColour}{the colour of the grass on the pitch}

\item{plotObject}{a ggplot object. Can just be a blank ggplot object too}
}
\description{
Adds pitch markings, green background, etc. for plot background to look like
a football pitch. ggplot adds things in order so initiate a ggplot object,
add the pitch marking with this function, and then add the data you want to
add
}
\examples{
ggplot() + geom_pitch()
}
